Created a basic recipe website.
Added a home page with three links to different recipes, Lasagna, Honey
Glazed Chicken, and Hollandaise Sauce.

Each link respectively has an image of the product, a description,
an unordered list of ingredients, and an ordered list of instructions to
make each respective recipe.

On the bottom of each recipe, I have added a link back to the home page.